Bug fixing and minor adjustments
The parser was not able to determine in which category it
should filter for a specific key. Since it misinterpreted the first
key and the duplicated values and completely ignored the
category, incorrect results were returned. This issue has now
been fixed. In addition, some parts of the code have been
refactored and adjustments have been made.